see question on first connect
